test(dedupe): consolidate position_detail tests (5 → 3 files)
Consolidate PositionDetailModal test files in tests/unit/gpt_trader/tui/widgets:
- test_position_detail_modal_init.py + recent_fills.py → test_position_detail.py
- test_position_detail_modal_calculations.py + fill_quality.py → test_position_detail_calculations.py
- test_position_detail_modal_trade_classification.py → test_position_detail_classification.py (rename)

Cluster: 4d680f449c31
Files: 5 → 3 (-2)
Tests: 16 passing
