+++ tests/unit/gpt_trader/tui/widgets/test_position_detail_classification.py
